workflow added
To set up a Maven build for each pull request merge into the master branch using GitHub Actions, you can create a workflow that triggers on the push event to the master branch. This will ensure that a build is performed every time a pull request is merged into the master branch.